fix : casting
암시적 형 변환을 명시적으로 변경합니다.

+++ Win32Game/NPC.cpp
@@ -61,7 +61,7 @@ void NPC::Render()
 			);
 		}
 	}
-	_TextBox->SetLocation(Vector3(startX+400, startY-50, 0));
+	_TextBox->SetLocation(Vector3((float)(startX + 400), (float)(startY - 50) , 0));
 	ComponentRender();
 }
